requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
discord==1.7.3
discord-components==1.0.0
discord.py==1.7.3
dnspython==2.1.0
h11==0.12.0
httpcore==0.12.3
httpx==0.16.1
idna==3.2
instant-api-client==0.1.2
multidict==5.1.0
numpy==1.21.0
Pillow==8.2.0
prsaw==0.4.0
pymongo==3.11.4
rfc3986==1.5.0
sniffio==1.2.0
typing-extensions==3.10.0.0
yarl==1.6.3